requirements.txt
numpy>=1.26.0
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
scipy>=1.12.0
python-dateutil>=2.9.0
pyyaml>=6.0.0
pytest>=8.0.0
